Update aliases for INPUT, FILTER, and OUTPUT resources
Update the alias attributes for most of the INPUT, FILTER, and OUTPUT
resources as these are used for the Prometheus metrics.

The labels for the metrics used in Prometheus use the alias name
exclusively, whereas the logs use the name and module for additional
context. Therefore, set the alias values to be more explicit in their
function to make it easier to read and review fluent-bit metrics.