Refactor the TSDB Series Alert in Prometheus
Refactor the monitoring of the increase in the series managed by the
TSDB in Prometheus such that it works of the noted new series from the
scrapes, and not the databases, and it alerts on a more sustained rate
of increase.

Specifically, it will monitor the increases over the last hour, and so
long as the increases exceed 1% per hour for more than three hours, only
then will it alert. This is as opposed to a single sudden large increase
which may be more expected.