Add support for a new notifications channel in Slack
Add support for a new notifications chennel in the Slack provider
configuration for Alertmanager which allows some Alerts to be sent there
where they do not need to trigger an action directly, but may mean
something needs adjusting.

This can include information about Pods exceeding the memory or CPU
requests as they are not exceeding the limits, and so won't be throttled
or terminated. However, typical usage is now above the guaranteed
amount, and the amount considered by the by the scheduler if the Pod is
moved to another host, creating the potential for unstable operation.