Refactor the kube_pod_info label merging for all Rules
Refactor the label merge for kube_pod_info to group by on only required
information, which should remove potential duplicate entries on the
left-hand-side of the group_left query. This will resolve when the
kube-state-metrics service is restarted due to Node or Deployment
upgrade which results in new values for some labels, including instance,
which breaks the mappings.