Disable the No New Documents Alert for ElasticSearch
Disable the individual node "No New Documents" Alert for ElasticSearch,
as its not unusual for a node not to be written to when only a few
shards are being processed. The "Critical" version of the alert still
monitors when no documents are written to the ElasticSearch cluster as a
whole, so it is still monitored.