Add global ignores for new false-positive Trivy checks
Add global ignores for KSV039 and KSV040 checks as these are known bugs
finding false-positive errors in unsupported resources. Additionally,
add overrides for KSV032, KSV033, KSV034, and KSV035, as these are all
related to the storage of contains in locally creaeted and managed
container registries, which I'm not currently intending to support.