Add additional PrometheusRules for ElasticSearch
Add additional PrometheusRules alerts for ElasticSearch which will
monitor the disk space on each of the nodes in the cluster based on the
current sizes of the indexes being stored.

The first check ensures that at least 28 days of logs and events can be
stored on the available PVC, and the second checks alert if there are
less than three and one days of space available.