Refactor the .trivyignore.yaml Configuration
Refactor the .trivyignore.yaml configuration file as the short format
for the misconfiguration IDs are no longer accepted by some checks
(namely AVD-KSV-0037 has begun to fail the checks on the files it is
marked to ignore). Additionally ensure that all overrides are in ID
order, and update the details of AVD-KSV-0037 to match the current
resources and check output.