Add the trivy-operator and the Dashboard to Clusters
Add the trivy-operator service to monitor the Kubernetes cluster and its
resources, as well as the trivy-operator Dashboard to allow it to
display and review the basic findings.

Additionally, automate the configuration of the Application and Provider
in Authentik to enable SSO support for the application. Access will only
be via Tailnet.