Migrate from Ingress to the Gateway API for Trivy
Refactor the configuration for the Gateway API, implementing Envoy
Gateway as the controller, with three sets of Gateways for external,
internal, and Tailnet traffic. Then, remove the Ingress configuration
for the Trivy Dashboard and set up new HTTPRoute and SecurityPolicy
resources for connecting to the Gateway with OIDC-based authentication.

This also requires updates for MetalLB to add the loadBalancerClass
argeument and external-dns to allow it to find and process the HTTPRoute
resources for assigning IP addresses.